Fix bug in gold standard splitter
Old version of gold standard splitter was discarding information from
PharmacotherapyDB because it was showing up in both the training and
holdout sets. The new version ensures that all of PharmacotherapyDB is
used, and also ensures that the training assumed negative examples do
not overlap at all with the holdout set's relations. However, some
possible relation combinations in the holdout set may appear in the
training data because they are part of the original PharmacotherapyDB.